requirements.txt
asgiref==3.8.1
Django==5.2.1
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
sqlparse==0.5.3
